Upgraded parsing exception reporting
Exceptions are now reported in a beginner-friendly way by printing the line where the error has occurred, color-coding the error range, adding arrows to the error range (for color-blind support), and then printing a simple, yet informative error message below. The collector and resolver will be upgraded in a subsequent commit.